refactor(code base): restructure code base and folders [Finishes #159191309]
