+++ server/test/Diary/postDiaryEntryTest/postDiaryEntryValidation/description/forbiddenSpecialCharacters.js
@@ -5,16 +5,16 @@ import server from '../../../../../app';
 chai.should();
 chai.use(chaiHttp);
 
-describe('diary entries', () => {
-  describe('PUT: /api/v1/entries/<entry Id>', () => {
-    it('it should respond with an error message if title field is empty', (done) => {
+describe('add diary entry', () => {
+  describe('POST: /api/v1/entries', () => {
+    it('it should respond with an error message if description field consists of forbidden special characters', (done) => {
       const diaryEntry = {
-        title: '',
-        description: 'update description 1',
+        title: 'A boy has no name',
+        description: 'Anthony is my name &',
         privacy: 'public'
       };
       chai.request(server)
-        .put('/api/v1/entries/1')
+        .post('/api/v1/entries')
         .send(diaryEntry)
         .end((err, res) => {
           res.should.have.status(400);
@@ -22,9 +22,9 @@ describe('diary entries', () => {
           res.body.should.be.a('object');
           res.body.should.have.property('errors');
           res.body.errors.should.be.a('object');
-          res.body.errors.should.have.property('title');
-          res.body.errors.title.should.be.a('string');
-          res.body.errors.title.should.eql('Title field is required');
+          res.body.errors.description.should.eql('Description must consist of alphanumeric \n'
+          + 'characters and special characters (? - , . ), and it must include at \n'
+          + 'least one alphabet');
           done();
         });
     });
