+++ server/test/Diary/postDiaryEntryTest/postDiaryEntryValidation/title/nonAlphanumericCharacter.js
@@ -4,16 +4,16 @@ import server from '../../../../../app';
 
 chai.should();
 chai.use(chaiHttp);
-describe('diary entries', () => {
-  describe('PUT: /api/v1/entries/<entry Id>', () => {
-    it('it should respond with an error message if title field has fewer than 4 characters', (done) => {
+describe('Add diary entries', () => {
+  describe('POST: /api/v1/entries', () => {
+    it('it should respond with an error message if title field has characters that aren\'t alphanumeric', (done) => {
       const diaryEntry = {
-        title: 'blo',
+        title: 'A boy ? no name',
         description: 'Boom',
         privacy: 'private'
       };
       chai.request(server)
-        .put('/api/v1/entries/1')
+        .post('/api/v1/entries')
         .send(diaryEntry)
         .end((err, res) => {
           res.should.have.status(400);
@@ -23,8 +23,7 @@ describe('diary entries', () => {
           res.body.errors.should.be.a('object');
           res.body.errors.should.have.property('title');
           res.body.errors.title.should.be.a('string');
-          res.body.errors.title.should.eql('Title must have a minimum length of 4 characters and a maximum \n'
-          + 'length of 70 characters');
+          res.body.errors.title.should.eql('Title should consist of only alphanumeric characters');
           done();
         });
     });
