+++ server/validators/postDiaryEntryValidator/index.js
@@ -1,0 +1,13 @@
+import postDiaryEntryPrivacyValidator from './privacyValidator';
+
+const postDiaryEntryValidator = (data) => {
+  const errors = postDiaryEntryPrivacyValidator(data);
+
+  return {
+    errors,
+    isValid: Object.keys(errors).length === 0
+  };
+};
+
+
+export default postDiaryEntryValidator;
